Update for v72 of the Platform SDK

+++ scripts/update_doc_classes_xml.sh
@@ -14,7 +14,9 @@ die() {
 $GODOT --doctool "$SCRIPT_DIR/.." --path "$PROJECT" --gdextension-docs --xr-mode off \
     || die "Failed to regenerate XML using Godot's --doctool"
 
-openxr_vendors_classes=$(cd "$SCRIPT_DIR/../thirdparty/godot_openxr_vendors/doc_classes" && ls *.xml)
-for x in $openxr_vendors_classes; do
-    rm -f "$SCRIPT_DIR/../doc_classes/$x"
-done
+if [ -n "$GODOT_OPENXR_VENDORS_SOURCE" ]; then
+    openxr_vendors_classes=$(cd "$GODOT_OPENXR_VENDORS_SOURCE/doc_classes" && ls *.xml)
+    for x in $openxr_vendors_classes; do
+        rm -f "$SCRIPT_DIR/../doc_classes/$x"
+    done
+fi
